Add a test (but for a simpler case)
