+++ plugin/src/constants.ts
@@ -1,0 +1,1 @@
+export const LAST_SHOPIFY_BULK_OPERATION = `LAST_SHOPIFY_BULK_OPERATION`;
