+++ plugin/types/interface.d.ts
@@ -32,6 +32,13 @@ interface BulkOperationNode {
   objectCount: string;
   url: string;
   id: string;
-  errorCode: "ACCESS_DENIED" | "INTERNAL_SERVER_ERROR" | "TIMEOUT";
+  errorCode?: "ACCESS_DENIED" | "INTERNAL_SERVER_ERROR" | "TIMEOUT";
   query: string;
+}
+
+interface CurrentBulkOperationResponse {
+  currentBulkOperation: {
+    id: string;
+    status: BulkOperationStatus;
+  };
 }
